Implementing "reading pages" for classes
There was a need for required reading pages which outline the
needed class texts and other materials that students must buy
for the class. This has been implemented.

Other cleanup and tidying was done as well, including some
cosmetic changes on the pages and the suppression of classes
which have no associated handouts/materials.

Added cookie use language to the About page.

Bumped the version number.