Cosmetic cleanup / lock refresh
There was some general cosmetic cleanup, but the most significant
change is that the refresh of the system has been made async by
running it in a thread. This allows the server to return control
back to Dropbox while refreshing the data continues.

This introduced a possible race condition where two threads might
be overlapping (and, truth be told, the race was likely already
there). This has been handled by including a lock at the start of
refresh, and an unlock at the end. Therefore if multiple threads
attempt to enter the refresh logic each one proceeds serially.