Change authentication
Make changes so that the auth no longer is driven by a URL in an
email sent to the student, but is rather verified with a 6 digit
code. This is much cleaner and allows more control in the browser.
It also mirrors what is being done on the Manuals website.