Notify instructors about 'contact' request
It used to be the case that the information from the contact
page only went to the #contact channel of the groundschool
Slack workspace. It was INTENDED to only be communications
about website problem, but was regularly mis-used to ask about
class related problems. So emailing the instructors has now
been added so everyone concerned is likely (hopefully) informed
of queries from students.