Guard against ADMIN being added as STUDENT
Ensure that the ADMIN collection is searched first so that if an
ADMIN user has (foolishly) been added as a STUDENT they still end
up being given admin privileges. Also check during STUDENT addition
that the email address isn't already registered as an ADMIN. Give
an error (personalized for George) if that happens. General cleanup
of unused <style> selectors.